src/ingestion/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
jsonschema==4.21.1
structlog==24.1.0
